Move API result structs to its own file, and leave the helper structs in separate file
